Use unsigned 16bit integers for locations
That allows maximum of 65535 columns and lines.

+++ src/lfortran/parser/location.h
@@ -1,6 +1,7 @@
 #ifndef LFORTRAN_PARSER_LOCATION_H
 #define LFORTRAN_PARSER_LOCATION_H
 
+#include <cstdint>
 #include <vector>
 #include <lfortran/ast.h>
 
@@ -9,10 +10,10 @@ namespace LFortran
 
 struct Location
 {
-  int first_line;
-  int first_column;
-  int last_line;
-  int last_column;
+  uint16_t first_line;
+  uint16_t first_column;
+  uint16_t last_line;
+  uint16_t last_column;
 };
 
 } // namespace LFortran
